DAS
An old algorithm, when I was starting to learn how to code. So, don't take this implementation too serious. This algorithm, was built using the library OpenCV, and the code have been writing in C++. The objective of this algorithm was to detected if appears an region of anommalia in the pacient  magnetic ressonance results.